+++ client/driver/docker.go
@@ -169,6 +169,7 @@ type DockerDriverConfig struct {
 	ForcePull        bool                `mapstructure:"force_pull"`         // Always force pull before running image, useful if your tags are mutable
 	MacAddress       string              `mapstructure:"mac_address"`        // Pin mac address to container
 	SecurityOpt      []string            `mapstructure:"security_opt"`       // Flags to pass directly to security-opt
+	MemLimitDisable  bool                `mapstructure:"mem_limit_disable"`  // Use soft memory limit instead of hard
 }
 
 // Validate validates a docker driver config
@@ -457,6 +458,9 @@ func (d *DockerDriver) Validate(config map[string]interface{}) error {
 			},
 			"security_opt": &fields.FieldSchema{
 				Type: fields.TypeArray,
+			},
+			"mem_limit_disable": &fields.FieldSchema{
+				Type: fields.TypeBool,
 			},
 		},
 	}
@@ -870,7 +874,15 @@ func (d *DockerDriver) createContainerConfig(ctx *ExecContext, task *structs.Tas
 		config.WorkingDir = driverConfig.WorkDir
 	}
 
-	memLimit := int64(task.Resources.MemoryMB) * 1024 * 1024
+	var memLimit, memReservation int64
+
+	taskMemReserved := int64(task.Resources.MemoryMB) * 1024 * 1024
+
+	if driverConfig.MemLimitDisable {
+		memReservation = taskMemReserved
+	} else {
+		memLimit = taskMemReserved
+	}
 
 	if len(driverConfig.Logging) == 0 {
 		if runtime.GOOS != "darwin" {
@@ -885,7 +897,9 @@ func (d *DockerDriver) createContainerConfig(ctx *ExecContext, task *structs.Tas
 
 	hostConfig := &docker.HostConfig{
 		// Convert MB to bytes. This is an absolute value.
-		Memory: memLimit,
+
+		Memory:            memLimit,
+		MemoryReservation: memReservation,
 		// Convert Mhz to shares. This is a relative value.
 		CPUShares: int64(task.Resources.CPU),